MyCFATool/requirements.txt
dash>=2.0.0
flask-login>=0.6.0
plotly>=5.0.0
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.25.0
pyyaml>=6.0
apscheduler>=3.9.0
reportlab>=3.6.0
redis==4.6.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
statsmodels>=0.13.0
pillow>=9.0.0
werkzeug>=2.0.0
flask>=2.0.0
locust>=2.0.0